Implement changes required by 64bit build

+++ xtc/Xtc.hh
@@ -6,6 +6,7 @@
 #include "Src.hh"
 
 #include <stdint.h>
+#include <stddef.h>
 
 namespace Pds {
 
@@ -21,8 +22,8 @@ namespace Pds {
       damage(_damage), src(_src), contains(_tag), extent(sizeof(Xtc)) {}
     Xtc(const TypeId& _tag, const Src& _src, const Damage& _damage) : damage(_damage), src(_src), contains(_tag), extent(sizeof(Xtc)) {}
     
-    void* operator new(unsigned size, char* p)     { return (void*)p; }
-    void* operator new(unsigned size, Xtc* p)      { return p->alloc(size); }
+    void* operator new(size_t size, char* p)     { return (void*)p; }
+    void* operator new(size_t size, Xtc* p)      { return p->alloc(size); }
     
     char*        payload()       const { return (char*)(this+1); }
     int          sizeofPayload() const { return extent - sizeof(Xtc); }
